pyproject.toml
[tool.poetry]
name = "stableswap-ng"
version = "0.1.0"
description = "Stableswap-ng environment for testing"
authors = ["Curve.fi"]
license = "MIT"
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
poetry = "1.5.1"
vyper = "0.3.10"
pycryptodome = "^3.18.0"
pre-commit = "^3.3.3"
titanoboa-zksync = {git = "https://github.com/DanielSchiavini/titanoboa-zksync.git", tag = "v0.1.2"} # inherits titanoboa 0.1.10

[tool.poetry.group.dev.dependencies]
black = "22.3.0"
flake8 = "4.0.1"
isort = "5.12.0"
mamushi = "^0.0.2a1"


[tool.poetry.group.testing.dependencies]
eip712 = {git = "https://github.com/DanielSchiavini/eip712.git", rev = "main"}
eth-account = "~0.13.0"
ipython = "^8.14.0"
hypothesis = "^6.79.3"
pytest = "^7.4.0"
pytest-xdist = "^3.3.1"
pytest-forked = "^1.6.0"
pytest-repeat = "^0.9.1"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
(
  /(
    \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
)
'''
line-length = 100
target_version = ['py310']

[tool.isort]
profile = "black"
py_version = 310
known_first_party = "poetry"

[tool.pytest.ini_options]
markers = [
  "only_plain_tokens", "only_oracle_tokens", "only_rebasing_tokens",
  "skip_plain_tokens", "skip_oracle_tokens", "skip_rebasing_tokens",
  "extensive_token_pairs",
  "only_basic_pool", "only_meta_pool",
  "skip_basic_pool", "skip_meta_pool",
]
filterwarnings = [
  "ignore:PytestUnknownMarkWarning"
]
